Структуризация csrfmw
